+++ aim/resource/utils.py
@@ -1,0 +1,2 @@
+def round10e5(val):
+    return round(val * 10e5) / 10e5